refactor python code structure
