+++ bigqueryClient.py
@@ -1,0 +1,9 @@
+from google.cloud import bigquery
+
+def get_users():
+    # Function to retrieve user data from BigQuery
+    pass
+
+def insert_posts(data):
+    # Function to insert data into the BigQuery 'tblInstagramData' table
+    pass
