+++ dataProcessor.py
@@ -1,0 +1,5 @@
+import pandas as pd
+
+def process_data(data):
+    # Function to clean and filter the data
+    pass
